requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.4
python-dotenv